requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
elasticsearch==7.17.4
idna==3.3
image-match @ git+https://github.com/EdjoLabs/image-match@4262f441c12bdd5ec07de449471e82df42126ccb
imageio==2.21.1
networkx==2.8.5
numpy==1.23.2
packaging==21.3
peewee==3.15.1
Pillow==9.2.0
pyparsing==3.0.9
pyTelegramBotAPI==4.7.0
python-dotenv==0.20.0
PyWavelets==1.3.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.0
six==1.16.0
soupsieve==2.3.2.post1
tifffile==2022.8.12
tqdm==4.64.0
urllib3==1.26.11
